API reference

PackageAnalyzer.analyze! — Method

analyze!(root, pkg_entries::AbstractVector{<:Tuple{PkgEntry, Union{VersionNumber, Symbol}}}; auth::GitHub.Authorization=github_auth(), sleep=0) -> Vector{Package}
analyze!(root, pkg_entries::AbstractVector{<:PkgEntry}; auth::GitHub.Authorization=github_auth(), sleep=0) -> Vector{Package}

Analyze all packages in the iterable pkg_entries, using threads, cloning them to root if a directory with their uuid does not already exist. Returns a Vector{Package}.

Optionally, use pairs (PkgEntry, Union{VersionNumber, Symbol}) to specify the version numbers, or pass a keyword argument version. The version number may be a VersionNumber, or :dev, or :stable. When pairs are passed, the keyword argument version will be ignored.

If the GitHub authentication is non-anonymous and the repository is on GitHub, the list of contributors to the repositories is also collected, after waiting for sleep seconds for each entry (useful to avoid getting rate-limited by GitHub). See PackageAnalyzer.github_auth to obtain a GitHub authentication.

PackageAnalyzer.analyze! — Method

analyze!(root, package::PkgEntry; auth::GitHub.Authorization=github_auth(), version::Union{VersionNumber, Symbol}=:dev) -> Package

Analyze the package whose entry in the registry is in the dir directory, cloning the package code to joinpath(root, uuid) where uuid is the UUID of the package, if such a directory does not already exist.

If the GitHub authentication is non-anonymous and the repository is on GitHub, the list of contributors to the repository is also collected, after waiting for sleep seconds. Only the number of contributors will be shown in the summary. See PackageAnalyzer.github_auth to obtain a GitHub authentication.

PackageAnalyzer.analyze — Method

analyze(name_or_dir_or_url::AbstractString; repo = "", reachable=true, subdir="", registry=general_registry(), auth::GitHub.Authorization=github_auth(), version::AbstractVersion=:dev)

Analyze the package pointed to by the mandatory argument and return a summary of its properties.

If name_or_dir_or_url is a valid Julia identifier, it is assumed to be the name of a package available in registry. The function then uses find_package to find its entry in the registry and analyze its content.

If name_or_dir_or_url is a filesystem path, analyze the package whose source code is located at name_or_dir_or_url. Optionally repo and reachable a boolean indicating whether or not the package is reachable online, since these can't be inferred from the source code. The subdir keyword arguments indicates the subdirectory of dir under which the Julia package can be found.

Otherwise, name_or_dir_or_url is assumed to be a URL. The repository is cloned to a temporary directory and analyzed.

If the GitHub authentication is non-anonymous and the repository is on GitHub, the list of contributors to the repository is also collected. Only the number of contributors will be shown in the summary. See PackageAnalyzer.github_auth to obtain a GitHub authentication.

Pass the keyword argument version to confgiure which version of the code is analyzed. Options:

:dev to use the latest code in the repository

:stable to use the latest released version of the code, or

pass a VersionNumber to analyze a particular version of the package.

If version !== :dev, only the code associated to that version of the package will be downloaded. That means for packages in subdirectories, top-level information (like CI scripts) may be unavailable.

Example

You can analyze a package just by its name, whether you have it installed locally or not:

julia> analyze("Pluto"; version=v"0.18.0")
Package Pluto:
  * repo: https://github.com/fonsp/Pluto.jl.git
  * uuid: c3e4b0f8-55cb-11ea-2926-15256bba5781
  * version: 0.18.0
  * is reachable: true
  * tree hash: db1306745717d127037c5697436b04cfb9d7b3dd
  * Julia code in `src`: 8337 lines
  * Julia code in `test`: 5448 lines (39.5% of `test` + `src`)
  * documention in `docs`: 0 lines (0.0% of `docs` + `src`)
  * documention in README: 118 lines
  * has license(s) in file: MIT
    * filename: LICENSE
    * OSI approved: true
  * has license(s) in Project.toml: MIT
    * OSI approved: true
  * has `docs/make.jl`: false
  * has `test/runtests.jl`: true
  * has continuous integration: true
    * GitHub Actions

PackageAnalyzer.analyze — Method

analyze(package::PkgEntry; auth::GitHub.Authorization=github_auth(), sleep=0, version::AbstractVersion=:dev) -> Package
analyze(packages::AbstractVector{<:PkgEntry}; auth::GitHub.Authorization=github_auth(), sleep=0, version::AbstractVersion=:dev) -> Vector{Package}

Analyzes a package or list of packages using the information in their directory in a registry by creating a temporary directory and calling analyze!, cleaning up the temporary directory afterwards.

If the GitHub authentication is non-anonymous and the repository is on GitHub, the list of contributors to the repository is also collected after waiting for sleep seconds (useful to avoid getting rate-limited by GitHub). Only the number of contributors will be shown in the summary. See PackageAnalyzer.github_auth to obtain a GitHub authentication.

Example

julia> analyze(find_package("BinaryBuilder"))
Package BinaryBuilder:
  * repo: https://github.com/JuliaPackaging/BinaryBuilder.jl.git
  * uuid: 12aac903-9f7c-5d81-afc2-d9565ea332ae
  * version: dev
  * is reachable: true
  * tree hash: 13335f33356c8df9899472634e02552fd6f99ce4
  * Julia code in `src`: 4994 lines
  * Julia code in `test`: 1795 lines (26.4% of `test` + `src`)
  * documention in `docs`: 1129 lines (18.4% of `docs` + `src`)
  * documention in README: 22 lines
  * has license(s) in file: MIT
    * filename: LICENSE.md
    * OSI approved: true
  * has `docs/make.jl`: true
  * has `test/runtests.jl`: true
  * has continuous integration: true
    * GitHub Actions
    * Azure Pipelines

PackageAnalyzer.analyze — Method

analyze(m::Module; kwargs...) -> Package

If you want to analyze a package which is already loaded in the current session, you can simply call analyze, which uses pkgdir to determine its source code:

julia> using DataFrames

julia> analyze(DataFrames)
Package DataFrames:
  * repo: 
  * uuid: a93c6f00-e57d-5684-b7b6-d8193f3e46c0
  * version: 0.0.0
  * is reachable: true
  * tree hash: db2a9cb664fcea7836da4b414c3278d71dd602d2
  * Julia code in `src`: 15628 lines
  * Julia code in `test`: 21089 lines (57.4% of `test` + `src`)
  * documention in `docs`: 6270 lines (28.6% of `docs` + `src`)
  * documention in README: 21 lines
  * has license(s) in file: MIT
    * filename: LICENSE.md
    * OSI approved: true
  * has `docs/make.jl`: true
  * has `test/runtests.jl`: true
  * has continuous integration: true
    * GitHub Actions

PackageAnalyzer.analyze_manifest — Method

analyze_manifest([path_to_manifest]; registries=reachable_registries(),
                 auth=github_auth(), sleep=0)

Convienence function to run find_packages_in_manifest then analyze on the results. Positional argument path_to_manifest defaults to joinpath(dirname(Base.active_project()), "Manifest.toml").

PackageAnalyzer.analyze_path! — Method

analyze_path!(dest::AbstractString, repo::AbstractString; name="", uuid=UUID(UInt128(0)), subdir="", auth=github_auth(), sleep=0, tree_hash=nothing) -> Package

Analyze the Julia package located at the URL given by repo by cloning it to dest and calling analyze_path(dest). If tree_hash !== nothing, only the code associated to that tree hash is placed into dest. That allows analyzing particular version numbers, but in the case of packages in subdirectories, it also means that top-level information (like CI workflows) is unavailable.

If the clone fails, it returns a Package with reachable=false. If a name and uuid are provided, these are used to populate the corresponding fields of the Package. If the clone succeeds, the name and uuid are taken instead from the Project.toml in the package itself, and the values passed here are ignored.

If the GitHub authentication auth is non-anonymous and the repository is on GitHub, the list of contributors to the repository is also collected, after waiting for sleep seconds for each entry. See PackageAnalyzer.github_auth to obtain a GitHub authentication.

PackageAnalyzer.analyze_path — Method

analyze_path(dir::AbstractString; repo = "", reachable=true, subdir="", auth::GitHub.Authorization=github_auth(), sleep=0, only_subdir=false, version=v"0") -> Package

Analyze the package whose source code is located at the local path dir. If the package's repository is hosted on GitHub and auth is a non-anonymous GitHub authentication, wait for sleep seconds before collecting the list of its contributors.

only_subdir indicates that while the package's code does live in a subdirectory of the repo, dir points only to that code and we do not have access to the top-level code. We still pass non-empty subdir in this case, to record the fact that the package does indeed live in a subdirectory.

Pass version to store the associated version number. Since this call only has access to files on disk, it does not know the associated version number in any registry.

PackageAnalyzer.find_package — Method

find_package(pkg; registry = general_registry()) -> RegistryEntry

Returns the RegistryEntry for the package pkg. The singular version of find_packages.

PackageAnalyzer.find_packages — Function

find_packages(; registry = general_registry()) -> Vector{RegistryEntry}
find_packages(names::AbstractString...; registry = general_registry()) -> Vector{RegistryEntry}
find_packages(names; registry = general_registry()) -> Vector{RegistryEntry}

Find all packages in the given registry (specified by the registry keyword argument), the General registry by default. Return a vector of RegistryEntry pointing to to the directories of each package in the registry.

Pass a list of package names as the first argument to return the paths corresponding to those packages, or individual package names as separate arguments.

PackageAnalyzer.general_registry — Method

general_registry() -> RegistryInstance

Return the RegistryInstance associated to the General registry.

PackageAnalyzer.github_auth — Function

PackageAnalyzer.github_auth(token::String="")

Obtain a GitHub authetication. Use the token argument if it is non-empty, otherwise use the GITHUB_TOKEN and GITHUB_AUTH environment variables, if set and of length 40. If all these methods fail, return an anonymous authentication.