API reference

AnalyzeRegistry.analyze — Method

analyze(dir::AbstractString; repo = "", reachable=true, name=nothing, uuid=nothing)

Analyze the package whose source code is located at dir. Optionally repo and reachable a boolean indicating whether or not the package is reachable online, since these can't be inferred from the source code. If name or uuid are nothing, the directories Project.toml is parsed to infer the package's name and UUID.

Example

julia> using DataFrames

julia> analyze(pkgdir(DataFrames))
Package DataFrames:
  * repo:
  * uuid: a93c6f00-e57d-5684-b7b6-d8193f3e46c0
  * is reachable: true
  * lines of Julia code in `src`: 15347
  * lines of Julia code in `test`: 15654
  * has license(s) in file: MIT
    * filename: LICENSE.md
    * OSI approved: true
  * has documentation: true
  * has tests: true
  * has continuous integration: true
    * GitHub Actions

AnalyzeRegistry.analyze_from_registry! — Method

analyze_from_registry!(root, dir::AbstractString) -> Package

Analyze the package whose entry in the registry is in the dir directory, cloning the package code to joinpath(root, uuid) where uuid is the UUID of the package, if such a directory does not already exist.

AnalyzeRegistry.analyze_from_registry! — Method

analyze_from_registry!(root, packages::AbstractVector{<:AbstractString}) -> Vector{Package}

Analyze all packages in the iterable packages, using threads, cloning them to root if a directory with their uuid does not already exist. Returns a Vector{Package}.

AnalyzeRegistry.analyze_from_registry — Method

analyze_from_registry(dir::AbstractString) -> Package
analyze_from_registry(packages::AbstractVector{<:AbstractString}) -> Vector{Package}

Analyzes a package or list of packages using the information in their directory in a registry by creating a temporary directory and calling analyze_from_registry!, cleaning up the temporary directory afterwards.

Example

julia> analyze_from_registry(joinpath(general_registry(), "B", "BinaryBuilder"))
Package BinaryBuilder:
  * repo: https://github.com/JuliaPackaging/BinaryBuilder.jl.git
  * uuid: 12aac903-9f7c-5d81-afc2-d9565ea332ae
  * is reachable: true
  * lines of Julia code in `src`: 4733
  * lines of Julia code in `test`: 1520
  * has license(s) in file: MIT
    * filename: LICENSE.md
    * OSI approved: true
  * has documentation: true
  * has tests: true
  * has continuous integration: true
    * GitHub Actions
    * Azure Pipelines

AnalyzeRegistry.find_package — Method

find_package(pkg; registry = general_registry()) -> String

Returns the path to the entry in registry for the package pkg. The singular version of find_packages.

AnalyzeRegistry.find_packages — Function

find_packages(; registry = general_registry()) -> Vector{String}
find_packages(names::AbstractString...; registry = general_registry()) -> Vector{String}
find_packages(names; registry = general_registry()) -> Vector{String}

Find all packages in the given registry (specified by the registry keyword argument), the General registry by default. Return a vector with the paths to the directories of each package in the registry.

Pass a list of package names as the first argument to return the paths corresponding to those packages, or individual package names as separate arguments.

AnalyzeRegistry.general_registry — Method

general_registry() -> String

Guess the path of the General registry.